it21-igordmitrow/Kursova/requirements.txt
anyio==4.0.0
appnope==0.1.3
APScheduler==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.3
Brotli==1.1.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
cryptography==41.0.4
DAWG-Python==0.7.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.6
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.18.1
Flask==3.0.0
Flask-Cors==4.0.0
fqdn==1.5.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
multipart==0.2.4
mutagen==1.47.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.0
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.1
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pycparser==2.21
pycryptodomex==3.19.0
pydub==0.25.1
pyglet==2.0.9
Pygments==2.16.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-telegram-bot==20.6
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
rapidfuzz==3.3.1
redis==5.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
Send2Trash==1.8.2
setuptools==68.2.2
ShazamAPI==0.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spotipy==2.23.0
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.17.1
TgCrypto==1.2.5
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
tzlocal==5.1
uri-template==1.3.0
urllib3==2.0.6
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
yt-dlp==2023.9.24
ytmusicapi==1.3.0
